% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{MLE}
\alias{MLE}
\title{MLE using R.}
\usage{
MLE(n, data, stheta, llm, is.log = TRUE)
}
\arguments{
\item{n}{the numbers of data}

\item{data}{which dim is n*p, p is the length of one sample}

\item{stheta}{the vector with length p, the initial value of optimization function}

\item{llm}{the function of likelihood of one sample and the parameter, input should be (parameter,sample)}

\item{is.log}{logical; if TRUE, the return values of llm are given as log}
}
\value{
the MLE estimation of parameter
}
\description{
MLE is a fundamental estimation method of parameter, the function is working in condition that sample is  independent identical distribution (iid)
}
\examples{
\dontrun{
llm1<- function(data,theta) return(dnorm(data,theta[1],theta[2]))
llm2<- function(data,theta) return(dnorm(data,theta[1],theta[2],log=TRUE))
MLE(10,X[1:10],c(-1,2),llm1,FALSE)
MLE(100,X,c(2,0.5),llm2,TRUE)
}
}
