% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksort.R
\name{quicksort}
\alias{quicksort}
\title{quicksort and count using R.}
\usage{
quicksort(x, is.count = FALSE, decreasing = FALSE)
}
\arguments{
\item{x}{the vector to be sorted}

\item{is.count}{logical; if TRUE, return the times of comparisons and exchanges}

\item{decreasing}{logical; the sort be increasing or decreasing}
}
\value{
if is.count=FALSE, the output is the result as a vector, else if is.count=TRUE, the output is a list with result ,compare and exchange.
}
\description{
quicksort is a efficiency but not a robust algorithm, the function mainly shows the exchange and comparison times of quicksort
}
\examples{
\dontrun{
X<-rnorm(10,0,1)
quicksort(X)
quicksort(X,is.count=TRUE)
}
}
